(PDOC-206) parser, handler, statement, code_object
Adds a parser for json, a task handler, a task statement and a task
code_object. the parser uses JSON.parse instead of an AST and some
finagling has to be done in the statement and the handler to get YARD to
be ok with this. When YARD is registering the task code_object, it
expects it to have things like a docstring and comments and so on, but
it doesn't have any of those things so they are faked.